services/notification-service/src/pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = ""
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiohttp (==3.11.18)",
    "gunicorn (==23.0.0)",
    "backoff (==2.2.1)",
    "async-fastapi-jwt-auth (==0.6.6)",
    "alembic (==1.15.2)",
    "python-dotenv (==1.1.0)",
    "sqlalchemy (==2.0.40)",
    "asyncpg (==0.30.0)",
    "psycopg (==3.2.9)",
    "uvicorn (==0.34.1)",
    "pydantic-settings (==2.9.1)",
    "fastapi[all] (==0.115.12)",
    "httpx (>=0.28.1,<0.29.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
